Collect guardrail validation duration and attach it to the guardrail events (#4119)
<!--
Thank you so much for your contribution!

Please fill in all the sections below.
Please open the PR as a draft initially. Once it is reviewed and
approved, we will ask you to add documentation and examples.
Please note that PRs with breaking changes or without tests will be
rejected.

Please note that PRs will be reviewed based on the priority of the
issues they address.
We ask for your patience. We are doing our best to review your PR as
quickly as possible.
Please refrain from pinging and asking when it will be reviewed. Thank
you for understanding!
-->

## Issue

Fix https://github.com/langchain4j/langchain4j/issues/4118

## Change

- Added Duration duration() method to GuardrailExecutedEvent interface
and builder
- Implemented duration field in DefaultGuardrailExecutedEvent
- Modified AbstractGuardrailExecutor to measure execution time using
System.nanoTime()
- Duration is automatically captured around guardrail validation and
included in fired events


## General checklist

- [X] There are no breaking changes (API, behaviour)
- [X] I have added unit and/or integration tests for my change
- [X] The tests cover both positive and negative cases
- [X] I have manually run all the unit and integration tests in the
module I have added/changed, and they are all green
- [X] I have manually run all the unit and integration tests in the
[core](https://github.com/langchain4j/langchain4j/tree/main/langchain4j-core)
and
[main](https://github.com/langchain4j/langchain4j/tree/main/langchain4j)
modules, and they are all green

Co-authored-by: Dmytro Liubarskyi <ljubarskij@gmail.com>