Implement all of the Gemini Files APIs (#4068)
## Issue
Closes #3949

## Change
This implements the complete Gemini Files API as documented here:
https://ai.google.dev/gemini-api/docs/files

***Additions***
* `GeminiFiles.uploadFile()` - Initiates a resumable upload that gets
the URL to upload the file to. After it's successful it uploads the file
to the provided URL and returns a `GeminiFile`.
* `GeminiFiles.getMetadata` - Retrieves meta data for the given file
name e.g. `files/upload.txt`. This is the same metadata as returned from
the uploadFile() call although the state might have changes.
* `GeminiFiles.listFiles()` - List all uploaded files. Files are
automatically deleted after 48 hours so this call will return all files
that were uploaded in the last 48 hours
* `GeminiFiles.deleteFile()` - Manually delete a file. This might be
useful as there's a project limit of 20GB

***TODO***:
 * Add examples / Docs
* Add File upload API for `GoogleAiGeminiBatchChatModel` (#3942 & #3945)

*HttpClient*: Unfortunately, I ended up using `java.net.http.HttpClient`
directly. So no wrapped `HttpClient`. I tried using the client that is
available but unfortunately the `body` is hard-coded to be always
`String`. As the file upload requires the body to be binary/bytes I
opted for using `java.net.http.HttpClient` directly. I had a look into
add/changing the `dev.langchain4j.http.client.HttpClient` but ultimately
this would a bunch of boilerplate to the response/request objects and
having two body methods (e.g. `body(String ...)` and `bodyBytes(...)`)
which is not great IMHO. Making the client, request and response objects
generic touched over 250 files, so I stayed away from that one
:sweat_smile:


## General checklist
- [x] There are no breaking changes (API, behaviour)
- [x] I have added unit and/or integration tests for my change
- [x] The tests cover both positive and negative cases
- [x] I have manually run all the unit and integration tests in the
module I have added/changed, and they are all green
- [x] I have manually run all the unit and integration tests in the
[core](https://github.com/langchain4j/langchain4j/tree/main/langchain4j-core)
and
[main](https://github.com/langchain4j/langchain4j/tree/main/langchain4j)
modules, and they are all green

---------

Co-authored-by: Dmytro Liubarskyi <ljubarskij@gmail.com>