Gemini (async) batch embed api (#4086)
## Issue
Progresses #3916 - This is one of the last PRs for the Gemini Batch API
support. The batch API calls using the Files API is still left to do

## Change
This introduces `GoogleAiGeminiBatchEmbeddingModel` which is similar to
`GoogleAiGeminiBatchChatModel` - as the batch API is actually the same
for both. I opted to split the `GoogleAiGeminiBatchChatModel` into two
and creating a generic `GeminiBatchProcessor` that does the actual
calling of the batch APIs. The latter deals with the logic of processing
the requests and wrapping the responses in a sealed `BatchResponse`. The
mapping between LangChain4j request/response objects and the Gemini API
request/response objects is done in each `GoogleAiGeminiBatch..Model`
through a `RequestPreparer<REQUEST, API_REQUEST, API_RESPONSE,
RESPONSE>`.

Sonar is giving me a hard time with the full word generics [REQUEST,
API_REQUEST, API_RESPONSE, RESPONSE] but IMHO it beats: [R1, R2, R3,
R4]. Let me know if you have any opinions on this.

Other than processor this PR introduces the
`GoogleAiGeminiBatchEmbeddingModel`which is exactly the same as the
preexisting `GoogleAiGeminiBatchChatModel` but with different [REQUEST,
API_REQUEST, API_RESPONSE, RESPONSE] generics.

Finally I choose to make `BatchRequestResponse` public as to expose the
`BatchRequestResponse` through where this was previously in
`GoogleAiGeminiBatchChatModel`. Meaning this is technically a breaking
change.


## General checklist
- [ ] There are no breaking changes (API, behaviour)
- [x] I have added unit and/or integration tests for my change
- [x] The tests cover both positive and negative cases
- [x] I have manually run all the unit and integration tests in the
module I have added/changed, and they are all green
- [x] I have manually run all the unit and integration tests in the
[core](https://github.com/langchain4j/langchain4j/tree/main/langchain4j-core)
and
[main](https://github.com/langchain4j/langchain4j/tree/main/langchain4j)
modules, and they are all green
<!-- Before adding documentation and example(s) (below), please wait
until the PR is reviewed and approved. -->
- [ ] I have added/updated the
[documentation](https://github.com/langchain4j/langchain4j/tree/main/docs/docs)
- [ ] I have added an example in the [examples
repo](https://github.com/langchain4j/langchain4j-examples) (only for
"big" features)

---------

Co-authored-by: Mario Fusco <mario.fusco@gmail.com>
Co-authored-by: Julio <52171109+wadajo@users.noreply.github.com>
Co-authored-by: Xin Wang <xinwang@apache.org>
Co-authored-by: Dmytro Liubarskyi <ljubarskij@gmail.com>
Co-authored-by: Óscar del Pozo <44649617+odelpozo-denodo@users.noreply.github.com>
Co-authored-by: Andrea Di Maio <85737936+andreadimaio@users.noreply.github.com>
Co-authored-by: Mathias Geat <mathias.geat@camunda.com>
Co-authored-by: Guillaume Laforge <glaforge@gmail.com>
Co-authored-by: Antonio Goncalves <antonio.goncalves@gmail.com>
Co-authored-by: Linn Aung <linnyehtutaung@gmail.com>
Co-authored-by: Hemanth Savasere <hemanth.savasere@gmail.com>
Co-authored-by: Oleksandr Klymenko <alexanderklmn@gmail.com>
Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>