Fill name support in PromptTemplate (#4180)
`PromptTemplateFactory.Input` defines the template and the name.
`PromptTemplate` creates input only with template and doesn't provide
the capability to define a template name.

This PR provides the capability to set the template name with a new
method

`public static PromptTemplate from(String template, String name)`

In my case I need to fill this template name for Quarkus LangChain4J
which defines promp template with Quarkus Qute syntax. I need this
template name to provide the capability to debug (with IDE breakpoint)
this Qute template.

## General checklist
 * [x]  There are no breaking changes (API, behaviour)
 * [ ]  I have added unit and/or integration tests for my change
 * [ ]  The tests cover both positive and negative cases
* [ ] I have manually run all the unit and integration tests in the
module I have added/changed, and they are all green
* [ ] I have manually run all the unit and integration tests in the
[core](https://github.com/langchain4j/langchain4j/tree/main/langchain4j-core)
and
[main](https://github.com/langchain4j/langchain4j/tree/main/langchain4j)
modules, and they are all green
* [ ] I have added/updated the
[documentation](https://github.com/langchain4j/langchain4j/tree/main/docs/docs)

---------

Signed-off-by: azerr <azerr@redhat.com>