Fix BatchSuccess mapping for Gemini BatchResponse (#4175)
## Issue

Closes #

## Change
Fixes 2 pretty glaring issues with the batch APIs. As the batching APIs
can take up to a day to complete there are no integration tests for it.
This resulted in me refactoring some of the methods on how the
successful responses are mapped and it going untested, breaking the
successful batches.

Jackson doesn't automatically map the responses and they need to be
manually mapped using `convertValue`:
```diff
- .map(BatchCreateResponse.InlinedResponseWrapper::response)
+ .map(wrapper -> objectMapper.convertValue(wrapper.response(), GeminiEmbeddingResponse.class))
```
And:
```diff
- .map(BatchCreateResponse.InlinedResponseWrapper::response)
+ .map(wrapper -> objectMapper.convertValue(wrapper.response(), GeminiGenerateContentResponse.class))
```

I've added tests for the (`BatchSuccess`) response serialization process
to ensure the changed code actually fixed it.


The other change is how unique `modelNames` are counted in the
batchResponse, earlier it was a requirement to set the `modelName` on a
`ChatRequest` which it isn't in the synchronous request. So now they're
all added to a set (including the `modelName` on the `ModelBuilder` and
make sure there's only a single unique modelName.

```java
var modelNames = Stream.concat(requests.stream().map(ChatRequest::modelName), Stream.of(modelName))
                .filter(Objects::nonNull)
                .collect(Collectors.toSet());
```

To finish it all up, I've also added examples here:
https://github.com/langchain4j/langchain4j-examples/pull/179

## General checklist
<!-- Please double-check the following points and mark them like this:
[X] -->
- [x] There are no breaking changes (API, behaviour)
- [x] I have added unit and/or integration tests for my change
- [x] The tests cover both positive and negative cases
- [x] I have manually run all the unit and integration tests in the
module I have added/changed, and they are all green
- [x] I have manually run all the unit and integration tests in the
[core]

---------

Co-authored-by: Dmytro Liubarskyi <ljubarskij@gmail.com>