feat: support structured outputs for Anthropic models (#4220)
<!--
Thank you so much for your contribution!

Please fill in all the sections below.
Please open the PR as a draft initially. Once it is reviewed and
approved, we will ask you to add documentation and examples.
Please note that PRs with breaking changes or without tests will be
rejected.

Please note that PRs will be reviewed based on the priority of the
issues they address.
We ask for your patience. We are doing our best to review your PR as
quickly as possible.
Please refrain from pinging and asking when it will be reviewed. Thank
you for understanding!
-->

## Issue
<!-- Please specify the ID of the issue this PR is addressing. For
example: "Closes #1234" or "Fixes #1234" -->
Closes #4174 

## Change
<!-- Please describe the changes you made. -->
Added structured output support for Anthropic models.

- Introduced `responseFormat()` and `supportedCapabilities()` methods
for `AnthropicChatModel` and `AnthropicStreamingChatModel`.
- Modeled JSON schemas in Anthropic requests via
`AnthropicResponseFormat`.
- Added a helper method in `AnthropicMapper` (`toAnthropicMap()`) to map
to the JSON schema format expected by the Anthropic API. Some features
(such as unions and complex types in enums) are still not supported, so
using the traditional `JsonSchemaUtils.toMap()` does not work for
certain edge cases.
- Removed `supportsJsonResponseFormat*()` methods from
`AnthropicChatModelIT` and `AnthropicStreamingChatModelIT`, and ensured
all tests pass, as requested in the issue description. Some tests had to
be overridden, since structured outputs are still a beta feature in
Anthropic so models require a special configuration.

## General checklist
<!-- Please double-check the following points and mark them like this:
[X] -->
- [X] There are no breaking changes (API, behaviour)
- [X] I have added unit and/or integration tests for my change
- [X] The tests cover both positive and negative cases
- [X] I have manually run all the unit and integration tests in the
module I have added/changed, and they are all green
- [X] I have manually run all the unit and integration tests in the
[core](https://github.com/langchain4j/langchain4j/tree/main/langchain4j-core)
and
[main](https://github.com/langchain4j/langchain4j/tree/main/langchain4j)
modules, and they are all green
<!-- Before adding documentation and example(s) (below), please wait
until the PR is reviewed and approved. -->
- [X] I have added/updated the
[documentation](https://github.com/langchain4j/langchain4j/tree/main/docs/docs)
- [ ] I have added an example in the [examples
repo](https://github.com/langchain4j/langchain4j-examples) (only for
"big" features)
- [ ] I have added/updated [Spring Boot
starter(s)](https://github.com/langchain4j/langchain4j-spring) (if
applicable)

---------

Co-authored-by: Dmytro Liubarskyi <ljubarskij@gmail.com>