Todo task/added additional date functions in sql filter parser class (#4269)
<!--
Thank you so much for your contribution!

Please fill in all the sections below.
Please open the PR as a draft initially. Once it is reviewed and
approved, we will ask you to add documentation and examples.
Please note that PRs with breaking changes or without tests will be
rejected.

Please note that PRs will be reviewed based on the priority of the
issues they address.
We ask for your patience. We are doing our best to review your PR as
quickly as possible.
Please refrain from pinging and asking when it will be reviewed. Thank
you for understanding!
-->

## Change
Completed a TODO task in SqlFilterParser class where I have added
additional Date functions (WEEKOFYEAR/DAYOFYEAR/DAYOFMONTH/DAYOFWEEK)
that will be converted to a Filter if these are present in a SQL Query.


## General checklist
<!-- Please double-check the following points and mark them like this:
[X] -->
- [X] There are no breaking changes (API, behaviour)
- [X] I have added unit and/or integration tests for my change
- [x] I have manually run all the unit and integration tests in the
module I have added/changed, and they are all green